Feat: 승인, 참가, 대기 제외한 마무리 작업 완료

+++ src/components/pages/postContent/Comments.vue
@@ -22,9 +22,7 @@
                   timeForToday(commentList.updatedAt)
                 }}</span></strong
               >
-              <p class="content">
-                {{ commentList.comment }}
-              </p>
+              <p v-html="putBr(commentList.comment)" class="content"></p>
             </div>
           </div>
           <div class="right">
@@ -47,49 +45,57 @@
             <button class="user__profile">
               <img :src="profileUrl" alt="유저 프로필" />
             </button>
-            <template v-if="isLogin">
-              <template v-if="checkReject() === true">
-                <p class="reject" type="text">
-                  참가가 거절되어 댓글을 다실 수 없어요
-                </p>
+            <div class="user__infos">
+              <strong class="nickname">
+                {{ loginedUserName }}
+              </strong>
+              <template v-if="isLogin">
+                <!-- FIXME: getCheckReject 작동안됨-->
+                <template v-if="getCheckReject === true">
+                  <p class="reject" type="text">
+                    참가가 거절되어 댓글을 다실 수 없어요
+                  </p>
+                </template>
+                <textarea
+                  v-else
+                  v-model="comment"
+                  @keyup.enter.exact="submitComment"
+                  ref="comment"
+                  class="add__comment"
+                  type="text"
+                  placeholder="댓글 달기"
+                ></textarea>
               </template>
-              <textarea
-                v-else
-                v-model="comment"
-                @keyup.enter.exact="submitComment"
-                ref="comment"
-                class="add__comment"
-                type="text"
-                placeholder="댓글 달기"
-              ></textarea>
-              <Button
-                v-if="isCommentLength"
-                @click="submitComment"
-                v-bind="{
-                  width: '80px',
-                  height: '40px',
-                  boxShadow: 'none',
-                }"
-                ><i class="material-icons">arrow_forward</i></Button
-              >
-              <Button
-                v-else
-                v-bind="{
-                  width: '80px',
-                  height: '40px',
-                  backgroundColor: '#CCCCCC',
-                  boxShadow: 'none',
-                }"
-                disabled
-                ><i class="material-icons">arrow_forward</i></Button
-              >
-            </template>
 
-            <div v-else class="logouted" @click="this.$router.push('/login')">
-              <button class="button--logouted" type="text">
-                로그인 해주세요
-              </button>
+              <div v-else class="logouted" @click="this.$router.push('/login')">
+                <button class="button--logouted" type="text">
+                  로그인 해주세요
+                </button>
+              </div>
             </div>
+          </div>
+          <div v-show="getCheckReject" class="right">
+            <Button
+              v-if="isCommentLength"
+              @click="submitComment"
+              v-bind="{
+                width: '80px',
+                height: '40px',
+                boxShadow: 'none',
+              }"
+              ><i class="material-icons">arrow_forward</i></Button
+            >
+            <Button
+              v-else
+              v-bind="{
+                width: '80px',
+                height: '40px',
+                backgroundColor: '#CCCCCC',
+                boxShadow: 'none',
+              }"
+              disabled
+              ><i class="material-icons">arrow_forward</i></Button
+            >
           </div>
         </li>
       </ul>
@@ -99,8 +105,8 @@
 
 <script>
 import { mapGetters } from 'vuex'
-import { getUserIdToCookie } from '~/utils/cookies'
-import { timeForToday } from '~/utils/function'
+import { getUserIdToCookie, getUserFromCookie } from '~/utils/cookies'
+import { timeForToday, putBr } from '~/utils/function'
 import { userDetailInfo } from '~/api'
 import { createComment, deleteComment } from '~/api/postContent'
 import TagArea from '~/components/pages/postContent/TagArea'
@@ -113,6 +119,7 @@ export default {
       comment: '',
       isSubmit: false,
       userId: '',
+      loginedUserName: '',
     }
   },
   components: {
@@ -144,6 +151,9 @@ export default {
     checkHost() {
       return this.author._id === this.userId
     },
+    getCheckReject() {
+      return this.checkReject()
+    },
   },
   methods: {
     async submitComment() {
@@ -164,15 +174,19 @@ export default {
       // TODO: \n을 바꿔줘야 함
     },
     async deleteComments(event) {
-      const li = event.target.closest('li')
-      const commentId = li.getAttribute('commentId')
-      const userData = {
-        id: commentId,
+      if (window.confirm('댓글을 삭제하시겠어요?')) {
+        const li = event.target.closest('li')
+        const commentId = li.getAttribute('commentId')
+        const userData = {
+          id: commentId,
+        }
+        console.log(userData)
+        const res = await deleteComment({ data: userData })
+        window.alert('댓글이 삭제되었어요')
+        this.$emit('rerender')
+      } else {
+        return
       }
-      console.log(userData)
-      const res = await deleteComment({ data: userData })
-      console.log(res, 'deleteComment')
-      this.$emit('rerender')
     },
 
     checkCommentor(event) {
@@ -181,21 +195,22 @@ export default {
       // // 댓글 작성자 _id, 로그인된 userId가 같은지 비교
       return commentorId === this.userId
     },
-
     async checkReject() {
       const res = await userDetailInfo(this.userId) // 댓작성자의 정보 불러옴
       const username = res.data.username.split('/')[1]
       const array = JSON.parse(username)
       const result = array.some(
-        // username: { postId, state }
         v => v.postid === this.postId && v.state === 'reject',
         // TODO: postId로 저장했는데 postid로 저장되는지 확인
       )
-      console.log(result, 'checkReject')
+      console.log(array, result)
       return result
     },
     getUserId() {
       this.userId = getUserIdToCookie()
+    },
+    getUserName() {
+      this.loginedUserName = getUserFromCookie()
     },
     doubleSubmitCHeck() {
       if (this.isSubmit) {
@@ -206,9 +221,10 @@ export default {
       }
     },
     timeForToday,
+    putBr,
   },
   created() {
-    this.getUserId()
+    this.getUserId(), this.getUserName()
   },
 }
 </script>
@@ -237,9 +253,18 @@ export default {
         border-bottom: 1px solid $COLOR_BORDER;
       }
       .left {
-        @include flexbox($jc: between);
+        @include flexbox($jc: start);
         &.etc {
           width: 100%;
+        }
+
+        .logouted {
+          width: 100%;
+
+          .button--logouted {
+            width: 100%;
+            text-align: left;
+          }
         }
 
         .reject {
@@ -254,9 +279,10 @@ export default {
         }
 
         .user__infos {
+          flex-grow: 1;
           .nickname {
             word-break: keep-all;
-            @include flexbox;
+            @include flexbox($jc: start);
             flex-grow: 1;
             flex-wrap: nowrap;
             margin-bottom: $INNER_PADDING_SMALL;
@@ -274,20 +300,20 @@ export default {
           .content {
             line-height: 1.3em;
           }
+
+          .add__comment {
+            font-family: inherit;
+            width: 100%;
+            border: none;
+            font-size: inherit;
+            resize: none;
+          }
         }
       }
 
       .right {
-        @include flexbox($jc: start);
+        @include flexbox($jc: between);
         flex-direction: column;
-      }
-
-      .add__comment {
-        font-family: inherit;
-        width: 100%;
-        border: none;
-        font-size: inherit;
-        resize: none;
       }
 
       .edit-area {
@@ -313,6 +339,44 @@ export default {
           width: 35px;
           height: 35px;
           object-fit: contain; // 일단은 contain으로 해놓음
+        }
+      }
+    }
+  }
+
+  @include responsive('sm') {
+    .comments {
+      &__header {
+        font-size: $FONT_L;
+      }
+
+      /* depth가 너무 길어짐 3depth 이상 들어가지 않게 이후 수정... */
+      &__list {
+        .item {
+          .left {
+            .user {
+              &__profile {
+                img {
+                  width: $SM_PROFILE_SIZE;
+                  height: $SM_PROFILE_SIZE;
+                }
+              }
+              &__infos {
+                .content,
+                .add__comment,
+                .reject {
+                  font-size: $FONT_S;
+                }
+              }
+            }
+          }
+
+          .right {
+            justify-content: center;
+            .button {
+              width: 50px !important;
+            }
+          }
         }
       }
     }
